Added Frienship between users
* Added 'My Friends' Link in nav
* Added 'add friend'/'delete friend' in communities and groups listing
* Can delete frienship from 'My friends'

+++ app/helpers/application_helper.rb
@@ -19,6 +19,7 @@ module ApplicationHelper
   end
 
   def link_to_friendship_action( user )
+    return "you" if user == current_user
     friend = current_user.friendship_for(user)
     if friend
       link_to 'delete friend', friend, :method => :delete, :confirm => "Are you sure?"
